requirements.txt
atlassian-python-api==3.2.0
click==7.1.2
pretty-tables==1.0.0
PyYAML==5.4.1
semantic-version==2.8.5
pprint36==3.9.0.2